src/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11
django-polymorphic==2.0.2
django-subdomains==2.1.0
idna==2.6
Pillow==5.0.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
django-easy-pdf==0.1.1
django-modeltranslation==0.12.2
django-settings-export==1.2.1
django-money==0.14.3
django_bot_crawler_blocker==1.0
mysql-connector-python==8.0.13